Support DOI in custom @norm bib type.
This also changes URLs to be displayed at the end of the entry (after
the DOI), if provided. Note that URLs should not be provided for norms,
except in cases where no DOI is available _and_ the URL points directly
to the document (i.e. PDF) file. Do not include URLs to a (potentially
changing) website!