Import 'siunitx' in document class
Closes #3.
This should rather be consider as a workaround than an actual fix.
Up to now the 'siunitx' package was imported inside the custom.sty.
This used to work with previous Latex versions.
In the current version this leads to build errors regarding the unicode representations
of omega and/or mu. Since the 'si' command itself does not raise an error it may not
be the actual reason for the build errors. Anyhow importing 'siunitx' inside the
emtthesis.cls leads to error free builds and correctly displayed symbols inside
every part of the document.
The configuration of the package ('sisetup') will still be done inside the custom.sty.