Do not use 'main' argument with babel
Babel enables two methods of setting the main language when working with
multiligual documents. When simply providing the languages as a list,
babel automatically selects the last entry as the main language.
Alternativly one can use the 'main=' argument. But when using 'main' the
hyperref package does not detect the language correctly and uses english
words e.g. for its 'autoref' function. Therefore, the main language is no
longer set via the argument, but is passed as the last list entry.